- libtecla/hash.c, libtecla/keytab.c: silenced compiler warnings    by casting chars to ints when using ctype.h (isspace & friends).

+++ hash.c
@@ -506,7 +506,7 @@ static HashNode *_new_HashNode(HashTable *hash, const char *name, int code,
     const char *src = name;
     char *dst = node->symbol.name;
     for( ; *src; src++,dst++)
-      *dst = tolower(*src);
+      *dst = tolower((int)*src);
     *dst = '\0';
   };
   return node;
